moved binning() to proc.py
